Refactor Json deserialization for test code only
Snapshot deserialization is only used for test code. Isolate
deserialization code from Moshi adapters to test helper class and
share it also for smoke tests
